requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pycryptodome>=3.18.0
torch>=2.0.0
flask>=2.3.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
